requirements.txt
requests==2.28.2
python-gitlab==3.13.0
openai==0.27.0
tiktoken==0.2.0
langchain==0.0.98
transformers==4.26.1
python-dotenv==1.0.0 